+++ packages/dom/src/state/gestures/types.ts
@@ -7,5 +7,9 @@ export interface StateHandlers {
 
 export interface Gesture {
   isActive: (options: Options) => void
-  subscribe: (element: Element, stateHandlers: StateHandlers) => () => void
+  subscribe: (
+    element: Element,
+    stateHandlers: StateHandlers,
+    options: Options
+  ) => () => void
 }