+++ packages/solid/src/types.ts
@@ -2,6 +2,7 @@ import type { JSX, ParentProps } from "solid-js"
 import type {
   ValueKeyframesDefinition,
   MotionKeyframesDefinition,
+  ViewOptions,
 } from "@motionone/dom"
 import type { MotionEvent, CustomPointerEvent, ViewEvent } from "@motionone/dom"
 import { AnimationOptions } from "@motionone/types"
@@ -46,6 +47,7 @@ export type Options = {
   press?: VariantDefinition
   exit?: VariantDefinition
   variants?: Record<string, Variant>
+  inViewOptions?: ViewOptions
   transition?: AnimationOptionsWithOverrides
 }