+++ packages/svelte/src/Motion.svelte
@@ -8,6 +8,7 @@
     MotionState,
     Variants,
     VariantDefinition,
+    InViewOptions,
   } from "@motionone/dom"
 
   const parentState = getContext<MotionState | undefined>(contextKey)
@@ -19,6 +20,7 @@
   export let hover: VariantDefinition | undefined = undefined
   export let press: VariantDefinition | undefined = undefined
   export let inView: VariantDefinition | undefined = undefined
+  export let inViewOptions: InViewOptions | undefined = undefined
   export let variants: Variants | undefined = undefined
   export let transition: AnimationOptionsWithOverrides | undefined = undefined
 
@@ -29,6 +31,7 @@
       hover,
       press,
       inView,
+      inViewOptions,
       variants,
       transition,
     },
@@ -46,6 +49,7 @@
       hover,
       press,
       inView,
+      inViewOptions,
       variants,
       transition,
     })